Refactored console progress via a new 'format_percentage' function and also provided access to this via the session variables '@progress_count', '@progress_total' and '@progress_value' along with adding a '-reload' option to the 'system_schedule' application protocol command.

+++ src/progress.h
@@ -35,8 +35,8 @@ struct console_progress : progress
    void output_progress(
     const std::string& message, unsigned long num = 0, unsigned long total = 0 );
 
-   int decimals;
-   int output_length;
+   size_t decimals;
+   size_t output_length;
 
    std::string output_prefix;
